Conflicts:
	code/modules/research/techweb/all_nodes.dm

+++ code/game/objects/items/implants/implant_stealth.dm
@@ -26,6 +26,10 @@
 
 /obj/structure/closet/cardboard/agent/open(mob/living/user, force = FALSE)
 	. = ..()
+
+	if(!.)
+		return
+
 	qdel(src)
 
 /obj/structure/closet/cardboard/agent/process()
